+++ src/main/java/net/spell_engine/client/particle/CustomizableParticleEffect.java
@@ -1,0 +1,22 @@
+package net.spell_engine.client.particle;
+
+import net.minecraft.particle.ParticleEffect;
+import net.spell_engine.client.util.Color;
+import org.jetbrains.annotations.Nullable;
+
+public interface CustomizableParticleEffect extends ParticleEffect {
+    class Appearance {
+        public @Nullable Color color;
+    }
+
+    void setAppearance(Appearance appearance);
+    Appearance getAppearance();
+
+    default Appearance createOrDefaultAppearance() {
+        var appearance = getAppearance() == null ? new Appearance() : getAppearance();
+        setAppearance(appearance);
+        return appearance;
+    }
+
+    CustomizableParticleEffect copy();
+}