+++ src/main/java/net/spell_engine/client/particle/GenericSparkParticle.java
@@ -1,4 +1,0 @@
-package net.spell_engine.client.particle;
-
-public class GenericSparkParticle {
-}