+++ src/main/java/net/spell_engine/fabric/client/FabricClientMod.java
@@ -96,6 +96,8 @@ public class FabricClientMod implements ClientModInitializer {
 
         ParticleFactoryRegistry.getInstance().register(SpellEngineParticles.fire_explosion.particleType(), SpellExplosionParticle.Factory::new);
 
+        ParticleFactoryRegistry.getInstance().register(SpellEngineParticles.spell_smoke.particleType(), SpellSmokeParticle.CosySmokeFactory::new);
+
         ModelLoadingPlugin.register(pluginCtx -> {
             pluginCtx.addModels(CustomModelRegistry.modelIds);
         });